research/requirements.txt
google-genai>=1.19.0 